+++ play-media-native/constants/event.ts
@@ -1,4 +1,4 @@
-import { IIndexable } from '../interfaces/indexable';
+import { CONTENT_TYPES, FIELD_TYPES } from './contentTypes';
 
 export const EVENT_INITIAL_STATE = {
   body: null,
@@ -14,10 +14,98 @@ export const EVENT_INITIAL_STATE = {
   similarEvents: [],
 };
 
-export const EVENT_FIELD_OVERRIDES: IIndexable = {
-  featuredImage: { single: true },
-  sport: { required: true, single: true },
-  title: { required: true },
+export const FIELD_OVERRIDES_EVENT = {
+  body: {
+    defaultValue: null,
+    required: false,
+    renderItem: null,
+    single: false,
+    type: FIELD_TYPES.RichText,
+    title: 'Body',
+  },
+  isFeatured: {
+    defaultValue: false,
+    required: false,
+    renderItem: null,
+    single: false,
+    type: FIELD_TYPES.Boolean,
+    title: 'Is featured?',
+  },
+  location: {
+    defaultValue: '',
+    required: false,
+    renderItem: null,
+    single: false,
+    type: FIELD_TYPES.ShortText,
+    title: 'Location',
+  },
+  title: {
+    defaultValue: '',
+    required: true,
+    renderItem: null,
+    single: false,
+    type: FIELD_TYPES.ShortText,
+    title: 'Title',
+  },
+  teaser: {
+    defaultValue: '',
+    required: false,
+    renderItem: null,
+    single: false,
+    type: FIELD_TYPES.ShortText,
+    title: 'Teaser',
+  },
+  timeAndDate: {
+    defaultValue: null,
+    required: false,
+    renderItem: null,
+    single: false,
+    type: FIELD_TYPES.Date,
+    title: 'Time & Date',
+  },
+  sport: {
+    defaultValue: null,
+    referenceType: CONTENT_TYPES.SPORT,
+    required: true,
+    renderItem: null,
+    single: true,
+    type: FIELD_TYPES.Reference,
+    title: 'Sport',
+  },
+  featuredImage: {
+    defaultValue: null,
+    required: false,
+    renderItem: null,
+    single: true,
+    type: FIELD_TYPES.Media,
+    title: 'Featured Image',
+  },
+  relatedMedia: {
+    defaultValue: [],
+    required: false,
+    renderItem: null,
+    single: false,
+    type: FIELD_TYPES.Media,
+    title: 'Related Media',
+  },
+  athletes: {
+    defaultValue: [],
+    referenceType: CONTENT_TYPES.ATHLETE,
+    required: false,
+    renderItem: null,
+    single: false,
+    type: FIELD_TYPES.Reference,
+    title: 'Athletes',
+  },
+  similarEvents: {
+    defaultValue: [],
+    referenceType: CONTENT_TYPES.EVENT,
+    required: false,
+    renderItem: null,
+    single: false,
+    type: FIELD_TYPES.Reference,
+    title: 'Related Events',
+  },
 };
 
 export const CREATE_EVENT_DISCARD_MESSAGE =
