+++ play-media-native/interfaces/contentItem/index.d.ts
@@ -18,3 +18,13 @@ export interface ContentItemUpdate {
     };
   };
 }
+
+export interface IFieldOverride {
+  defaultValue: any;
+  referenceType?: string;
+  required: boolean;
+  renderItem: ReactNode;
+  single: boolean;
+  title: string;
+  type: string;
+}
