+++ play-media-native/interfaces/indexable/index.ts
@@ -1,15 +1,3 @@
-import { ReactNode } from 'react';
-
 export interface IIndexable {
   [key: string]: any;
-}
-
-export interface IFieldOverride {
-  defaultValue: any;
-  referenceType?: string;
-  required: boolean;
-  renderItem: ReactNode;
-  single: boolean;
-  title: string;
-  type: string;
 }
