+++ play-media-native/interfaces/sport/index.ts
@@ -1,13 +1,23 @@
+import { Media } from '../media';
+
+export interface SportResponse {
+  id: string;
+  title: string;
+  description: string;
+  featuredImage: { results: Media[] };
+}
+
 export interface Sport {
   id: string;
   title: string;
   description: string;
+  featuredImage: Media;
 }
 
 export interface AllSportsResponse {
   data: {
     allSport: {
-      results: Sport[];
+      results: SportResponse[];
     };
   };
 }
